bajaj hackathon 22/backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
requests
pypdf
langchain
langchain-groq
langchain_community
sentence-transformers
faiss-cpu